Rename mock-data folder as mock
